refactor: 팝업 컴포넌트 분리
