+++ components/pages/detailPage/PostData/index.tsx
@@ -1,13 +1,6 @@
 import { useState } from "react";
-import customAxios from "apis/customAxios";
-import { useRouter } from "next/router";
 import Link from "next/link";
-import {
-  BiUserCircle,
-  BiDotsHorizontalRounded,
-  BiEditAlt,
-  BiTrash,
-} from "react-icons/bi";
+import { BiUserCircle, BiDotsHorizontalRounded } from "react-icons/bi";
 import {
   PostTitle,
   Row,
@@ -16,11 +9,9 @@ import {
   Column,
   SpanStyled,
   OptionBox,
-  OptionPopupWrapper,
-  OptionBtn,
-  Container,
 } from "@/detailComps/PostData/styled";
 import { useAppSelector } from "store/hooks";
+import Popup from "@/detailComps/Popup";
 
 interface PostDataProps {
   postId: number;
@@ -30,10 +21,6 @@ interface PostDataProps {
   likeNum: number;
   writerId: number;
   writer: string;
-}
-
-interface OptionPopupProps {
-  postId: number;
 }
 
 export default function PostData({
@@ -81,7 +68,7 @@ export default function PostData({
         <div>
           <span>좋아요 </span>
           <SpanStyled>{likeNum}</SpanStyled>
-          {popupOn && <OptionPopup postId={postId} />}
+          {popupOn && <Popup postId={postId} />}
         </div>
         {/* 로그인한 유저가 작성자가 아니면 수정 팝업 안보임 */}
         {+userId === writerId && (
@@ -92,55 +79,5 @@ export default function PostData({
       </Row>
       <hr />
     </div>
-  );
-}
-
-function OptionPopup({ postId }: OptionPopupProps) {
-  const router = useRouter();
-  const postCategory = router.pathname.split("/")[1];
-
-  // 게시글 삭제
-  const deletePost = async (id: number) => {
-    let url = "";
-    if (postCategory === "recruits") url += "/recruit-board/" + id;
-    if (postCategory === "projects") url += "/free-boards/" + id;
-
-    try {
-      const res = await customAxios.delete(url);
-      if (res.status === 200) {
-        await window.alert("게시물이 성공적으로 삭제되었습니다");
-        await router.push(`/${postCategory}`);
-      }
-    } catch (err) {
-      console.log(err);
-      window.alert("게시물 삭제에 실패했습니다");
-    }
-  };
-
-  const handleDelete = async () => {
-    if (window.confirm("정말 삭제하시겠습니까?")) {
-      await deletePost(postId);
-    }
-  };
-
-  const handleEdit = async () => {
-    await router.push(`/${postCategory}/edit/${postId}`);
-  };
-
-  return (
-    <OptionPopupWrapper>
-      <OptionBtn type="button" onClick={handleEdit}>
-        <Container>
-          <BiEditAlt size={20} />
-          <p>수정</p>
-        </Container>
-      </OptionBtn>
-      <OptionBtn type="button" onClick={handleDelete}>
-        <Container>
-          <BiTrash size={20} />
-          <p>삭제</p>
-        </Container>
-      </OptionBtn>
-    </OptionPopupWrapper>
   );
 }
